[idea] #0000 Refactored RefreshAction class
allowing 4 pmd violations
