+++ plugins/idea/kodebeagleidea/src/main/java/com/imaginea/kodebeagle/ui/ProjectTree.java
@@ -17,7 +17,7 @@
 
 package com.imaginea.kodebeagle.ui;
 
-import com.imaginea.kodebeagle.util.WindowEditorOps;
+import com.imaginea.kodebeagle.tasks.FetchFileContentTask;
 import com.imaginea.kodebeagle.model.CodeInfo;
 import com.imaginea.kodebeagle.object.WindowObjects;
 import com.imaginea.kodebeagle.util.ESUtils;
@@ -27,12 +27,8 @@ import com.intellij.ide.BrowserUtil;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.editor.EditorFactory;
 import com.intellij.openapi.fileEditor.FileEditorManager;
-import com.intellij.openapi.progress.PerformInBackgroundOption;
-import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.progress.ProgressManager;
-import com.intellij.openapi.progress.Task;
 import com.intellij.openapi.vfs.VirtualFile;
-import org.jetbrains.annotations.NotNull;
 
 import java.awt.Component;
 import java.awt.event.ActionEvent;
@@ -63,7 +59,6 @@ import javax.swing.tree.TreeNode;
 public class ProjectTree {
     private static final String OPEN_IN_NEW_TAB = "Open in New Tab";
     private WindowObjects windowObjects = WindowObjects.getInstance();
-    private WindowEditorOps windowEditorOps = new WindowEditorOps();
     private ESUtils esUtils = new ESUtils();
     private JSONUtils jsonUtils = new JSONUtils();
     private EditorDocOps editorDocOps = new EditorDocOps();
@@ -84,19 +79,11 @@ public class ProjectTree {
                         windowObjects.getjTree().getLastSelectedPathComponent();
                 if (selectedNode != null && selectedNode.isLeaf() && root.getChildCount() > 0) {
                     final CodeInfo codeInfo = (CodeInfo) selectedNode.getUserObject();
-                    ProgressManager.getInstance().run(new FetchFileContentTask(codeInfo));
+                    ProgressManager.getInstance().run(new FetchFileContentTask(
+                            windowObjects.getProject(), codeInfo));
                 }
             }
         };
-    }
-
-    private void updateMainPanePreviewEditor(final List<Integer> lineNumbers,
-                                             final String fileContents) {
-        final Document mainPanePreviewEditorDocument =
-                windowObjects.getWindowEditor().getDocument();
-        String contentsInLines =
-                editorDocOps.getContentsInLines(fileContents, lineNumbers);
-        windowEditorOps.writeToDocument(contentsInLines, mainPanePreviewEditorDocument);
     }
 
     public final Map<String, ArrayList<CodeInfo>> updateProjectNodes(
@@ -216,37 +203,6 @@ public class ProjectTree {
                 }
             }
         };
-    }
-
-    private class FetchFileContentTask extends Task.Backgroundable {
-
-        private static final String KODE_BEAGLE = "KodeBeagle";
-        private static final String FETCHING_FILE_CONTENT = "Fetching file content ...";
-        private final CodeInfo codeInfo;
-        private String fileContents;
-
-        public FetchFileContentTask(final CodeInfo pCodeInfo) {
-            super(windowObjects.getProject(),
-                    KODE_BEAGLE, true, PerformInBackgroundOption.ALWAYS_BACKGROUND);
-            this.codeInfo = pCodeInfo;
-        }
-
-        @Override
-        public void run(@NotNull final ProgressIndicator indicator) {
-            indicator.setText(FETCHING_FILE_CONTENT);
-            indicator.setFraction(0.0);
-            String fileName = codeInfo.getAbsoluteFileName();
-            fileContents = esUtils.getContentsForFile(fileName);
-
-            //Setting contents so that we can use that for Open in New Tab
-            codeInfo.setContents(fileContents);
-            indicator.setFraction(1.0);
-        }
-
-        @Override
-        public void onSuccess() {
-            updateMainPanePreviewEditor(codeInfo.getLineNumbers(), fileContents);
-        }
     }
 
     public final JTreeCellRenderer getJTreeCellRenderer() {
