+++ plugins/idea/kodebeagleidea/src/main/java/com/imaginea/kodebeagle/util/Utils.java
@@ -19,6 +19,7 @@ package com.imaginea.kodebeagle.util;
 
 import com.imaginea.kodebeagle.ui.KBNotification;
 import com.intellij.openapi.util.io.FileUtil;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileOutputStream;
@@ -32,7 +33,6 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import org.apache.commons.codec.binary.Hex;
 import org.jetbrains.annotations.NotNull;
-
 
 public final class Utils {
